requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
idna==3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
slack-cleaner2==3.2.0
slack-sdk==3.18.3
urllib3==1.26.12
